#402 Rename i18n.entry to i18n.message and add default formatting binding
