+++ hartshorn-i18n/src/main/java/org/dockbox/hartshorn/i18n/ResourceServiceImpl.java
@@ -23,7 +23,7 @@ import org.dockbox.hartshorn.di.annotations.inject.Binds;
 import org.dockbox.hartshorn.di.context.ApplicationContext;
 import org.dockbox.hartshorn.i18n.common.Language;
 import org.dockbox.hartshorn.i18n.common.Message;
-import org.dockbox.hartshorn.i18n.entry.MessageTemplate;
+import org.dockbox.hartshorn.i18n.message.MessageTemplate;
 import org.dockbox.hartshorn.util.HartshornUtils;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
