+++ hartshorn-i18n/src/main/java/org/dockbox/hartshorn/i18n/common/ResourceEntryModel.java
@@ -18,7 +18,7 @@
 package org.dockbox.hartshorn.i18n.common;
 
 import org.dockbox.hartshorn.di.context.ApplicationContext;
-import org.dockbox.hartshorn.i18n.entry.MessageTemplate;
+import org.dockbox.hartshorn.i18n.message.MessageTemplate;
 import org.dockbox.hartshorn.persistence.PersistentModel;
 
 import lombok.AllArgsConstructor;
