+++ hartshorn-i18n/src/main/java/org/dockbox/hartshorn/i18n/message/ConfigurableMessageFormatting.java
@@ -1,11 +1,13 @@
-package org.dockbox.hartshorn.i18n.entry;
+package org.dockbox.hartshorn.i18n.message;
 
 import org.dockbox.hartshorn.config.annotations.Value;
+import org.dockbox.hartshorn.di.annotations.inject.Binds;
 import org.dockbox.hartshorn.di.annotations.service.Service;
 
 import lombok.Getter;
 
 @Service
+@Binds(MessageFormatting.class)
 public class ConfigurableMessageFormatting implements MessageFormatting {
 
     @Getter @Value("hartshorn.i18n.primary")
