+++ hartshorn-i18n/src/main/java/org/dockbox/hartshorn/i18n/message/DefaultResources.java
@@ -15,7 +15,7 @@
  * along with this library. If not, see {@literal<http://www.gnu.org/licenses/>}.
  */
 
-package org.dockbox.hartshorn.i18n.entry;
+package org.dockbox.hartshorn.i18n.message;
 
 import org.dockbox.hartshorn.boot.Hartshorn;
 import org.dockbox.hartshorn.di.annotations.service.Service;
