+++ hartshorn-i18n/src/main/java/org/dockbox/hartshorn/i18n/message/MessageFormatting.java
@@ -15,7 +15,7 @@
  * along with this library. If not, see {@literal<http://www.gnu.org/licenses/>}.
  */
 
-package org.dockbox.hartshorn.i18n.entry;
+package org.dockbox.hartshorn.i18n.message;
 
 public interface MessageFormatting {
     String primary();
