+++ hartshorn-i18n/src/test/java/org/dockbox/hartshorn/i18n/MessageTemplateServiceTests.java
@@ -21,7 +21,7 @@ import org.dockbox.hartshorn.api.domain.Exceptional;
 import org.dockbox.hartshorn.di.context.ApplicationContext;
 import org.dockbox.hartshorn.i18n.common.Language;
 import org.dockbox.hartshorn.i18n.common.Message;
-import org.dockbox.hartshorn.i18n.entry.MessageTemplate;
+import org.dockbox.hartshorn.i18n.message.MessageTemplate;
 import org.dockbox.hartshorn.test.ApplicationAwareTest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
